# docs/openapi.yaml
openapi: 3.0.3
info:
  title: UAV Docking Management System API
  description: |
    Comprehensive REST API for managing UAV operations, real-time tracking, geofencing, and fleet management.
    
    ## Authentication
    The API uses role-based authentication with three user roles:
    - **USER**: Read-only access
    - **OPERATOR**: Can perform UAV operations and updates  
    - **ADMIN**: Full system access including delete operations
    
    ## Rate Limiting
    - Default: 100 requests per minute
    - Admin: 1000 requests per minute
    - API Key: 2000 requests per minute
  version: 1.0.0
  contact:
    name: UAV Management System Team
    email: support@uavmanagement.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.uavmanagement.com
    description: Production server
security:
  - sessionAuth: []
  - apiKeyAuth: []
paths:
...
  /api/location/update/{uavId}:
    post:
      tags:
        - Location Tracking
      summary: Update UAV location
      description: Update the current location of a UAV
      security:
        - sessionAuth: []
      parameters:
        - name: uavId
          in: path
          required: true
          schema:
            type: integer
          description: UAV ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LocationUpdateRequest'
      responses:
        '200':
          description: Location updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/location/current:
    get:
      tags:
        - Location Tracking
      summary: Get current locations
      description: Get current locations of all UAVs
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Current locations retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/LocationHistory'
  /api/location/history/{uavId}:
    get:
      tags:
        - Location Tracking
      summary: Get location history
      description: Get historical location data for a specific UAV
      security:
        - sessionAuth: []
      parameters:
        - name: uavId
          in: path
          required: true
          schema:
            type: integer
          description: UAV ID
        - name: startDate
          in: query
          schema:
            type: string
            format: date-time
          description: Start date (ISO 8601 format)
        - name: endDate
          in: query
          schema:
            type: string
            format: date-time
          description: End date (ISO 8601 format)
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
          description: Maximum number of records
      responses:
        '200':
          description: Location history retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/LocationHistory'
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    LocationUpdateRequest:
      type: object
      required:
        - latitude
        - longitude
      properties:
        latitude:
          type: number
          format: double
          example: 40.7589
        longitude:
          type: number
          format: double
          example: -73.9851
        altitude:
          type: number
          format: double
          example: 85.5
        speed:
          type: number
          format: double
          example: 25.3
        heading:
          type: number
          format: double
          example: 180.0
        batteryLevel:
          type: integer
          minimum: 0
          maximum: 100
          example: 75
        timestamp:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
    LocationHistory:
      type: object
      properties:
        id:
          type: integer
          example: 1
        uavId:
          type: integer
          example: 1
        timestamp:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        latitude:
          type: number
          format: double
          example: 40.7589
        longitude:
          type: number
          format: double
          example: -73.9851
        altitudeMeters:
          type: number
          format: double
          example: 85.5
        speedKmh:
          type: number
          format: double
          example: 25.3
        headingDegrees:
          type: number
          format: double
          example: 180.0
        batteryLevel:
          type: integer
          example: 75
        locationSource:
          type: string
          enum: [GPS, CELLULAR, WIFI, MANUAL, ESTIMATED, RADAR]
          example: "GPS"
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation completed successfully"
        timestamp:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: "UAV_NOT_FOUND"
            message:
              type: string
              example: "UAV with specified ID not found"
            details:
              type: object
        timestamp:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
  responses:
    BadRequest:
      description: Bad request - invalid input parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...